requirements.txt
openai
python-dotenv
rumps
py2app